requirements.txt
hydra-core==1.3.2
mlflow==2.19.0
scikit-learn==1.6.0
pandas==2.2.3
numpy==2.2.1
pytest==8.3.4
pytest-cov==6.0.0
